requirements.txt
asgiref==3.8.1
Django==5.0.3
django-extensions==3.2.3
django-hashid-field==3.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
factory-boy==3.3.0
Faker==24.3.0
hashids==1.3.1
PyJWT==2.8.0
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.4.4
